+++ flask_admin/contrib/pymongo/filters.py
@@ -35,9 +35,7 @@ class BasePyMongoFilter(filters.BaseFilter):
         :param data_type:
             Client data type
         """
-        super().__init__(column, name, options, data_type, url_value=url_value)
-
-        self.column = column
+        super().__init__(name, options, data_type, column=column, url_value=url_value)
 
 
 # Common filters
